+++ R/utils-qqassign.R
@@ -1,0 +1,11 @@
+#' Pipe operator
+#'
+#' See \code{rlang::\link[rlang]{:=}} for details.
+#'
+#' @name :=
+#' @rdname quasi-quote-assign
+#' @keywords internal
+#' @export
+#' @importFrom rlang :=
+#' @usage x := y
+NULL